Move registry::{Fallbacks, Api, Profile} to top level module
